move create list button to userlists
