+++ src/hooks/useWatchLists.js
@@ -1,4 +1,5 @@
 import { useState, useContext, useEffect } from 'react';
+import { useHistory } from 'react-router-dom';
 import { FirebaseContext } from '../firebase';
 
 function useWatchLists() {
@@ -6,9 +7,14 @@ function useWatchLists() {
   const [watchLists, setWatchLists] = useState([]);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState(null);
+  const history = useHistory();
 
   function createWatchList(newList) {
-    firebase.db.doc(`users/${user.uid}`).collection('lists').add(newList);
+    if (!user) {
+      history.push('/login');
+    } else {
+      firebase.db.doc(`users/${user.uid}`).collection('lists').add(newList);
+    }
   }
 
   useEffect(() => {
